requirements.txt
requests==2.32.5
beautifulsoup4==4.14.2
python-dotenv==1.2.1
playwright==1.55.0
psycopg2-binary==2.9.10
python-telegram-bot==21.10
